useComboboxFieldProps hook to unify handleBlur and handleChange functions for combobox formik fields
