+++ src/common/components/formFields/umbrellaEventSelectorField/UmbrellaEventSelectorField.tsx
@@ -1,18 +1,16 @@
-import { FieldProps, useField } from 'formik';
+import { FieldProps } from 'formik';
 import React from 'react';
-import { useTranslation } from 'react-i18next';
 
 import { getEventFields } from '../../../../domain/event/utils';
 import { EVENT_SORT_OPTIONS } from '../../../../domain/events/constants';
 import { EventFieldsFragment } from '../../../../generated/graphql';
 import { Language, OptionType } from '../../../../types';
-import getValue from '../../../../utils/getValue';
-import { getErrorText } from '../../../../utils/validationUtils';
 import EventSelector, {
   EventSelectorProps,
 } from '../../eventSelector/EventSelector';
+import useComboboxFieldProps from '../hooks/useComboboxFieldProps';
 
-type Props = EventSelectorProps & FieldProps;
+type Props = EventSelectorProps & FieldProps<string>;
 
 const getEventOption = (
   event: EventFieldsFragment,
@@ -28,18 +26,12 @@ const UmbrellaEventSelectorField: React.FC<Props> = ({
   helper,
   ...rest
 }) => {
-  const { t } = useTranslation();
-  const [, { touched, error }] = useField(name);
-
-  const errorText = getErrorText(error, touched, t);
-
-  const handleBlur = () => {
-    onBlur({ target: { id: name, value } });
-  };
-
-  const handleChange = (selected: OptionType | null) => {
-    onChange({ target: { id: name, value: getValue(selected?.value, null) } });
-  };
+  const { errorText, handleBlur, handleChange } = useComboboxFieldProps({
+    name,
+    onBlur,
+    onChange,
+    value,
+  });
 
   return (
     <EventSelector
